move WiFi settings to Kconfig for mpd_status example
Instead of requiring the user to create a wifi_config.h file, this will use the recommended way to add user-specific configurations from ESP-IDF.

At the same time this has the benefit, that the example even compiles when the user did not set the WiFi settings. The settings simply default to empty strings.